% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DvLink.R
\name{DvLink}
\alias{DvLink}
\title{MLHIM 2.5.0 DvLink}
\usage{
DvLink(pcm)
}
\arguments{
\item{pcm}{- the XML fragment to parse.}
}
\value{
A dataframe consisting of the vectors listed in the Description.
}
\description{
Returns a data.frame of the collected nodes of a DvLink restriction from the XML fragment.
}
\details{
The vectors are:

label - the descriptive name of the fragment.

vtb - valid time begin, the datetime that this data became valid

vte - valid time end, the datetime this data expired or ceased to be valid

ev_name - exceptional value name. When all data is valid according to the restrictions this is NA.
          See the MLHIM documentation for explainations of Exceptional Values.

link - A required URI. A dereferenceable or not, URL to another CCD or just the ccd-uuid.

relation - A required term, normally drawn from a vocabulary or ontology such as the OBO RO.

relation_uri - A URI where the definition of the relation element term can be found.
               Normally points to an ontology such as the OBO RO http://purl.obolibrary.org/obo/ro.owl
}
\examples{
parseBladderTestingOnly <- function(sourceName){

doc <- XML::xmlTreeParse(sourceName, handlers=list('comment'=function(x,...){NULL}), asTree = TRUE)
root <- XML::xmlRoot(doc)

pcm <- XML::getNodeSet(root, '//mlhim2:pcs-780f834c-f29f-46aa-bb50-3384e9884138', c(mlhim2='http://www.mlhim.org/ns/mlhim2/'))

data <- lapply(pcm, mlhim250rm::DvLink)
data <- data.table::rbindlist(data)
data$ccd <- XML::xmlName(root)
data$pcs <- 'pcs-780f834c-f29f-46aa-bb50-3384e9884138'
data$sourceName <- sourceName

return(data)
}
}

