% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DvOrdinal.R
\name{DvOrdinal}
\alias{DvOrdinal}
\title{MLHIM 2.5.0 DvOrdinal}
\usage{
DvOrdinal(pcm)
}
\arguments{
\item{pcm}{- the XML fragment to parse.}
}
\value{
A dataframe consisting of the vectors listed in the Description.
}
\description{
Returns a data.frame of the collected nodes of a DvOrdinal restriction from the XML fragment.
}
\details{
The vectors are:

label - the descriptive name of the fragment.

vtb - valid time begin, the datetime that this data became valid

vte - valid time end, the datetime this data expired or ceased to be valid

ev_name - exceptional value name. When all data is valid according to the restrictions this is NA.
          See the MLHIM documentation for explainations of Exceptional Values.

reference_ranges -

normal_status - Optional normal status indicator of value with respect to normal range for this value.
                Often included by lab, even if the normal range itself is not included.
                Coded by ordinals in series HHH, HH, H, (nothing), L, LL, LLL, etc.

ordinal - Value in ordered enumeration of values. The base integer is zero with any number of integer values used to order the symbols.
          Example 1: 0 = Trace, 1 = +, 2 = ++, 3 = +++, etc. Example 2: 0 = Mild, 1 = Moderate, 2 = Severe

symbol - Coded textual representation of this value in the enumeration, which may be strings made from “+” symbols, or other enumerations of
         terms such as “mild”, “moderate”, “severe”, or even the same number series as the values, e.g. “1”, “2”, “3”.
}
\examples{
parseBladderTestingOnly <- function(sourceName){

doc <- XML::xmlTreeParse(sourceName, handlers=list('comment'=function(x,...){NULL}), asTree = TRUE)
root <- XML::xmlRoot(doc)

pcm <- XML::getNodeSet(root, '//mlhim2:pcs-780f834c-f29f-46aa-bb50-3384e9884138', c(mlhim2='http://www.mlhim.org/ns/mlhim2/'))

data <- lapply(pcm, mlhim250rm::DvOrdinal)
data <- data.table::rbindlist(data)
data$ccd <- XML::xmlName(root)
data$pcs <- 'pcs-780f834c-f29f-46aa-bb50-3384e9884138'
data$sourceName <- sourceName

return(data)
}
}

