% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DvQuantity.R
\name{DvQuantity}
\alias{DvQuantity}
\title{MLHIM 2.5.0 DvQuantity}
\usage{
DvQuantity(pcm)
}
\arguments{
\item{pcm}{- the XML fragment to parse.}
}
\value{
A dataframe consisting of the vectors listed in the Description.
}
\description{
Returns a data.frame of the collected nodes of a DvQuantity restriction from the XML fragment.
}
\details{
The vectors are:

label - the descriptive name of the fragment.

vtb - valid time begin, the datetime that this data became valid

vte - valid time end, the datetime this data expired or ceased to be valid

ev_name - exceptional value name. When all data is valid according to the restrictions this is NA.
          See the MLHIM documentation for explainations of Exceptional Values.

reference_ranges -

normal_status - Optional normal status indicator of value with respect to normal range for this value.
                Often included by lab, even if the normal range itself is not included.
                Coded by ordinals in series HHH, HH, H, (nothing), L, LL, LLL, etc.

dvquantity_value - Decimal value of the measurement.

units_value - The name or type of the quantity. Examples are "kg/m2", “mmHg", "ms-1", "km/h". May or may not come from a standard terminology.

units_label - A group name for the enumeration set used for the units value.

magnitude_status - Optional status of magnitude. See the MLHIM 2.5.0 documentation.

error - Error margin of measurement, indicating error in the recording method or instrument (+/- %).
        A logical value of 0 indicates 100% accuracy, i.e. no error.

accuracy - Accuracy of the value in the magnitude attribute in the range 0% to (+/-)100%.
           A value of 0 means that the accuracy is unknown.
}
\examples{
parseBladderTestingOnly <- function(sourceName){

doc <- XML::xmlTreeParse(sourceName, handlers=list('comment'=function(x,...){NULL}), asTree = TRUE)
root <- XML::xmlRoot(doc)

pcm <- XML::getNodeSet(root, '//mlhim2:pcs-780f834c-f29f-46aa-bb50-3384e9884138', c(mlhim2='http://www.mlhim.org/ns/mlhim2/'))

data <- lapply(pcm, mlhim250rm::DvQuantity)
data <- data.table::rbindlist(data)
data$ccd <- XML::xmlName(root)
data$pcs <- 'pcs-780f834c-f29f-46aa-bb50-3384e9884138'
data$sourceName <- sourceName

return(data)
}
}

