% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DvTemporal.R
\name{DvTemporal}
\alias{DvTemporal}
\title{MLHIM 2.5.0 DvTemporal}
\usage{
DvTemporal(pcm)
}
\arguments{
\item{pcm}{- the XML fragment to parse.}
}
\value{
A dataframe consisting of the vectors listed in the Description.
}
\description{
Returns a data.frame of the collected nodes of a DvTemporal restriction from the XML fragment.
}
\details{
The vectors are:

label - the descriptive name of the fragment.

vtb - valid time begin, the datetime that this data became valid

vte - valid time end, the datetime this data expired or ceased to be valid

ev_name - exceptional value name. When all data is valid according to the restrictions this is NA.
          See the MLHIM documentation for explainations of Exceptional Values.

reference_ranges -

normal_status - Optional normal status indicator of value with respect to normal range for this value.
                Often included by lab, even if the normal range itself is not included.
                Coded by ordinals in series HHH, HH, H, nothing, L, LL, LLL, etc.

dvtemporal_date

dvtemporal_time

dvtemporal_datetime

dvtemporal_datetime_stamp

dvtemporal_day

dvtemporal_month

dvtemporal_year

dvtemporal_year_month

dvtemporal_month_day

dvtemporal_duration

dvtemporal_ymduration

dvtemporal_dtduration
}
\examples{
parseBladderTestingOnly <- function(sourceName){

doc <- XML::xmlTreeParse(sourceName, handlers=list('comment'=function(x,...){NULL}), asTree = TRUE)
root <- XML::xmlRoot(doc)

pcm <- XML::getNodeSet(root, '//mlhim2:pcs-780f834c-f29f-46aa-bb50-3384e9884138', c(mlhim2='http://www.mlhim.org/ns/mlhim2/'))

data <- lapply(pcm, mlhim250rm::DvTemporal)
data <- data.table::rbindlist(data)
data$ccd <- XML::xmlName(root)
data$pcs <- 'pcs-780f834c-f29f-46aa-bb50-3384e9884138'
data$sourceName <- sourceName

return(data)
}
}

